combine AvatarSystem into other avatar systems, fix bug with height not calculating properly
