+++ packages/engine/src/avatar/AvatarAnimationSystem.ts
@@ -1,17 +1,34 @@
 import { useEffect } from 'react'
-import { Bone, MathUtils, Vector3 } from 'three'
+import { Bone, MathUtils, Object3D, Vector3 } from 'three'
 
 import { insertionSort } from '@xrengine/common/src/utils/insertionSort'
-import { defineState, getState, startReactor, useHookstate } from '@xrengine/hyperflux'
+import {
+  createActionQueue,
+  defineState,
+  dispatchAction,
+  getState,
+  startReactor,
+  useHookstate
+} from '@xrengine/hyperflux'
 
 import { Axis } from '../common/constants/Axis3D'
 import { V_000 } from '../common/constants/MathConstants'
+import { isClient } from '../common/functions/isClient'
 import { isHMD } from '../common/functions/isMobile'
-import { EngineState } from '../ecs/classes/EngineState'
+import { proxifyQuaternion, proxifyVector3 } from '../common/proxies/createThreejsProxy'
 import { Entity } from '../ecs/classes/Entity'
 import { World } from '../ecs/classes/World'
-import { defineQuery, getComponent, getOptionalComponent, removeQuery } from '../ecs/functions/ComponentFunctions'
+import {
+  defineQuery,
+  getComponent,
+  getOptionalComponent,
+  hasComponent,
+  removeComponent,
+  removeQuery,
+  setComponent
+} from '../ecs/functions/ComponentFunctions'
 import { createPriorityQueue } from '../ecs/PriorityQueue'
+import { WorldNetworkAction } from '../networking/functions/WorldNetworkAction'
 import { RigidBodyComponent } from '../physics/components/RigidBodyComponent'
 import { VisibleComponent } from '../scene/components/VisibleComponent'
 import {
@@ -20,6 +37,7 @@ import {
   FrustumCullCameraComponent
 } from '../transform/components/DistanceComponents'
 import { updateGroupChildren } from '../transform/systems/TransformSystem'
+import { getCameraMode } from '../xr/XRState'
 import { updateAnimationGraph } from './animation/AnimationGraph'
 import { solveHipHeight } from './animation/HipIKSolver'
 import { solveLookIK } from './animation/LookAtIKSolver'
@@ -29,7 +47,11 @@ import { AnimationComponent } from './components/AnimationComponent'
 import { AvatarAnimationComponent, AvatarRigComponent } from './components/AvatarAnimationComponent'
 import { AvatarArmsTwistCorrectionComponent } from './components/AvatarArmsTwistCorrectionComponent'
 import { AvatarControllerComponent } from './components/AvatarControllerComponent'
-import { AvatarLeftHandIKComponent, AvatarRightHandIKComponent } from './components/AvatarIKComponents'
+import {
+  AvatarIKTargetsComponent,
+  AvatarLeftHandIKComponent,
+  AvatarRightHandIKComponent
+} from './components/AvatarIKComponents'
 import { AvatarHeadDecapComponent } from './components/AvatarIKComponents'
 import { AvatarHeadIKComponent } from './components/AvatarIKComponents'
 import { LoopAnimationComponent } from './components/LoopAnimationComponent'
@@ -44,6 +66,109 @@ export const AvatarAnimationState = defineState({
 
 const _vector3 = new Vector3()
 const _vec = new Vector3()
+
+/**
+ * Setup head-ik for entity
+ * @param entity
+ * @returns
+ */
+export function setupHeadIK(entity: Entity) {
+  const target = new Object3D()
+  target.name = `ik-head-target-${entity}`
+
+  setComponent(entity, AvatarHeadIKComponent, {
+    target,
+    rotationClamp: 0.785398
+  })
+
+  const headIK = getComponent(entity, AvatarHeadIKComponent)
+  proxifyVector3(AvatarHeadIKComponent.target.position, entity, headIK.target.position)
+  proxifyQuaternion(AvatarHeadIKComponent.target.quaternion, entity, headIK.target.quaternion)
+}
+
+export function setupLeftHandIK(entity: Entity) {
+  const leftHint = new Object3D()
+  leftHint.name = `ik-left-hint-${entity}`
+  const leftOffset = new Object3D()
+  leftOffset.name = `ik-left-offset-${entity}`
+  leftOffset.updateMatrix()
+  leftOffset.updateMatrixWorld(true)
+
+  const rig = getComponent(entity, AvatarRigComponent)
+
+  leftOffset.rotation.set(-Math.PI * 0.5, Math.PI, 0)
+
+  if (isClient) {
+    rig.rig.LeftShoulder.getWorldPosition(leftHint.position)
+    rig.rig.LeftArm.getWorldPosition(_vec)
+    _vec.subVectors(_vec, leftHint.position).normalize()
+    leftHint.position.add(_vec)
+    rig.rig.LeftShoulder.attach(leftHint)
+    leftHint.updateMatrix()
+    leftHint.updateMatrixWorld(true)
+  }
+
+  const target = new Object3D()
+  target.name = `ik-right-target-${entity}`
+
+  setComponent(entity, AvatarLeftHandIKComponent, {
+    target,
+    hint: leftHint,
+    targetOffset: leftOffset,
+    targetPosWeight: 1,
+    targetRotWeight: 1,
+    hintWeight: 1
+  })
+
+  const lefthand = getComponent(entity, AvatarLeftHandIKComponent)
+  proxifyVector3(AvatarLeftHandIKComponent.target.position, entity, lefthand.target.position)
+  proxifyQuaternion(AvatarLeftHandIKComponent.target.quaternion, entity, lefthand.target.quaternion)
+}
+// setComponent(entity, AvatarArmsTwistCorrectionComponent, {
+//   LeftHandBindRotationInv: new Quaternion(),
+//   LeftArmTwistAmount: 0.6,
+//   RightHandBindRotationInv: new Quaternion(),
+//   RightArmTwistAmount: 0.6
+// })
+
+export function setupRightHandIK(entity: Entity) {
+  const rightHint = new Object3D()
+  rightHint.name = `ik-right-hint-${entity}`
+  const rightOffset = new Object3D()
+  rightOffset.name = `ik-right-offset-${entity}`
+
+  const rig = getComponent(entity, AvatarRigComponent)
+
+  rightOffset.rotation.set(-Math.PI * 0.5, 0, 0)
+  rightOffset.updateMatrix()
+  rightOffset.updateMatrixWorld(true)
+
+  if (isClient) {
+    rig.rig.RightShoulder.getWorldPosition(rightHint.position)
+    rig.rig.RightArm.getWorldPosition(_vec)
+    _vec.subVectors(_vec, rightHint.position).normalize()
+    rightHint.position.add(_vec)
+    rig.rig.RightShoulder.attach(rightHint)
+    rightHint.updateMatrix()
+    rightHint.updateMatrixWorld(true)
+  }
+
+  const target = new Object3D()
+  target.name = `ik-right-target-${entity}`
+
+  setComponent(entity, AvatarRightHandIKComponent, {
+    target,
+    hint: rightHint,
+    targetOffset: rightOffset,
+    targetPosWeight: 1,
+    targetRotWeight: 1,
+    hintWeight: 1
+  })
+
+  const rightHand = getComponent(entity, AvatarRightHandIKComponent)
+  proxifyVector3(AvatarRightHandIKComponent.target.position, entity, rightHand.target.position)
+  proxifyQuaternion(AvatarRightHandIKComponent.target.quaternion, entity, rightHand.target.quaternion)
+}
 
 export default async function AvatarAnimationSystem(world: World) {
   await AnimationManager.instance.loadDefaultAnimations()
@@ -66,6 +191,9 @@ export default async function AvatarAnimationSystem(world: World) {
     AvatarRigComponent
   ])
   const avatarAnimationQuery = defineQuery([AnimationComponent, AvatarAnimationComponent, AvatarRigComponent])
+  const avatarIKTargetsQuery = defineQuery([AvatarIKTargetsComponent, AvatarRigComponent])
+
+  const avatarIKTargetsActionQueue = createActionQueue(WorldNetworkAction.avatarIKTargets.matches)
 
   const reactor = startReactor(() => {
     const state = useHookstate(getState(AvatarAnimationState))
@@ -97,7 +225,52 @@ export default async function AvatarAnimationSystem(world: World) {
   let sortedTransformEntities = [] as Entity[]
 
   const execute = () => {
-    const { elapsedSeconds, deltaSeconds, localClientEntity } = world
+    const { elapsedSeconds, deltaSeconds, localClientEntity, inputSources } = world
+
+    if (localClientEntity && hasComponent(localClientEntity, AvatarIKTargetsComponent)) {
+      const ikTargets = getComponent(localClientEntity, AvatarIKTargetsComponent)
+      const sources = Array.from(inputSources.values())
+      const head = getCameraMode() === 'attached'
+      const leftHand = !!sources.find((s) => s.handedness === 'left')
+      const rightHand = !!sources.find((s) => s.handedness === 'right')
+
+      const changed = ikTargets.head !== head || ikTargets.leftHand !== leftHand || ikTargets.rightHand !== rightHand
+
+      if (changed) dispatchAction(WorldNetworkAction.avatarIKTargets({ head, leftHand, rightHand }))
+    }
+
+    for (const action of avatarIKTargetsActionQueue()) {
+      const entity = world.getUserAvatarEntity(action.$from)
+      const targets = getComponent(entity, AvatarIKTargetsComponent)
+
+      targets.head = action.head
+      targets.leftHand = action.leftHand
+      targets.rightHand = action.rightHand
+    }
+
+    /** Add & remove IK Targets based on active target data */
+    for (const entity of avatarIKTargetsQuery()) {
+      const targets = getComponent(entity, AvatarIKTargetsComponent)
+
+      if (targets.head && !hasComponent(entity, AvatarHeadIKComponent)) setupHeadIK(entity)
+      if (!targets.head && hasComponent(entity, AvatarHeadIKComponent)) removeComponent(entity, AvatarHeadIKComponent)
+
+      if (targets.leftHand && !hasComponent(entity, AvatarLeftHandIKComponent)) setupLeftHandIK(entity)
+      if (!targets.leftHand && hasComponent(entity, AvatarLeftHandIKComponent)) {
+        const leftHand = getComponent(entity, AvatarLeftHandIKComponent)
+        leftHand?.hint?.removeFromParent()
+        removeComponent(entity, AvatarLeftHandIKComponent)
+      }
+
+      if (targets.rightHand && !hasComponent(entity, AvatarRightHandIKComponent)) setupRightHandIK(entity)
+      if (!targets.rightHand && hasComponent(entity, AvatarRightHandIKComponent)) {
+        const rightHand = getComponent(entity, AvatarRightHandIKComponent)
+        rightHand?.hint?.removeFromParent()
+        removeComponent(entity, AvatarRightHandIKComponent)
+      }
+    }
+
+    if (!isClient) return
 
     /**
      * 1 - Sort & apply avatar priority queue
@@ -235,9 +408,6 @@ export default async function AvatarAnimationSystem(world: World) {
 
       // Arms should not be straight for the solver to work properly
       // TODO: Make this configurable
-
-      // TODO: should we break hand IK apart into left and right components?
-      // some devices only support one hand controller. How do we handle that?
       // how do we report that tracking is lost or still pending?
       // FOR NOW: we'll assume that we don't have tracking if the target is at exactly (0, 0, 0);
       // we may want to add a flag for this in the future, or to generally allow animations to play even if tracking is available
@@ -286,38 +456,6 @@ export default async function AvatarAnimationSystem(world: World) {
         )
       }
     }
-
-    // for (const entity of armsTwistCorrectionQuery.enter()) {
-    //   const { bindRig } = getComponent(entity, AvatarRigComponent)
-    //   const twistCorrection = getComponent(entity, AvatarArmsTwistCorrectionComponent)
-    //   twistCorrection.LeftHandBindRotationInv.copy(bindRig.LeftHand.quaternion).invert()
-    //   twistCorrection.RightHandBindRotationInv.copy(bindRig.RightHand.quaternion).invert()
-    // }
-
-    // for (const entity of armsTwistCorrectionQuery()) {
-    //   const { rig, bindRig } = getComponent(entity, AvatarRigComponent)
-    //   const twistCorrection = getComponent(entity, AvatarArmsTwistCorrectionComponent)
-
-    //   if (rig.LeftForeArmTwist) {
-    //     applyBoneTwist(
-    //       twistCorrection.LeftHandBindRotationInv,
-    //       rig.LeftHand.quaternion,
-    //       bindRig.LeftForeArmTwist.quaternion,
-    //       rig.LeftForeArmTwist.quaternion,
-    //       twistCorrection.LeftArmTwistAmount
-    //     )
-    //   }
-
-    //   if (rig.RightForeArmTwist) {
-    //     applyBoneTwist(
-    //       twistCorrection.RightHandBindRotationInv,
-    //       rig.RightHand.quaternion,
-    //       bindRig.RightForeArmTwist.quaternion,
-    //       rig.RightForeArmTwist.quaternion,
-    //       twistCorrection.RightArmTwistAmount
-    //     )
-    //   }
-    // }
 
     /**
      * Since the scene does not automatically update the matricies for all objects,which updates bones,
