+++ packages/engine/src/avatar/AvatarCommonModule.ts
@@ -1,6 +1,5 @@
 import { SystemUpdateType } from '../ecs/functions/SystemUpdateType'
 import AvatarSpawnSystem from './AvatarSpawnSystem'
-import AvatarSystem from './AvatarSystem'
 
 export function AvatarCommonModule() {
   return [
@@ -8,11 +7,6 @@ export function AvatarCommonModule() {
       uuid: 'xre.engine.AvatarSpawnSystem',
       type: SystemUpdateType.UPDATE,
       systemLoader: () => Promise.resolve({ default: AvatarSpawnSystem })
-    },
-    {
-      uuid: 'xre.engine.AvatarSystem',
-      type: SystemUpdateType.UPDATE,
-      systemLoader: () => Promise.resolve({ default: AvatarSystem })
     }
   ]
 }
