+++ packages/engine/src/avatar/functions/avatarFunctions.ts
@@ -26,6 +26,7 @@ import { addObjectToGroup, GroupComponent, removeObjectFromGroup } from '../../s
 import { UpdatableCallback, UpdatableComponent } from '../../scene/components/UpdatableComponent'
 import { ObjectLayers } from '../../scene/constants/ObjectLayers'
 import { setObjectLayers } from '../../scene/functions/setObjectLayers'
+import { computeTransformMatrix, updateGroupChildren } from '../../transform/systems/TransformSystem'
 import { createAvatarAnimationGraph } from '../animation/AvatarAnimationGraph'
 import { applySkeletonPose, isSkeletonInTPose, makeTPose } from '../animation/avatarPose'
 import { retargetSkeleton, syncModelSkeletons } from '../animation/retargetSkeleton'
@@ -119,8 +120,13 @@ export const setupAvatarForUser = (entity: Entity, model: Object3D) => {
   if (avatar.model) removeObjectFromGroup(entity, avatar.model)
 
   setupAvatarModel(entity)(model)
-
   addObjectToGroup(entity, model)
+
+  computeTransformMatrix(entity)
+  updateGroupChildren(entity)
+
+  setupAvatarHeight(entity, model)
+
   setObjectLayers(model, ObjectLayers.Avatar)
   avatar.model = model
 }
@@ -171,8 +177,6 @@ export const rigAvatarModel = (entity: Entity) => (model: Object3D) => {
   syncModelSkeletons(model, targetSkeleton)
 
   setComponent(entity, AvatarRigComponent, { rig, bindRig: avatarBoneMatching(SkeletonUtils.clone(rootBone)) })
-
-  setupAvatarHeight(entity, model)
 
   const sourceHips = sourceSkeleton.bones[0]
   avatarAnimationComponent.rootYRatio = rig.Hips.position.y / sourceHips.position.y
